Removed search by date for now.
