+++ data/elasticsearch/seed-data.json
@@ -1,2 +1,2 @@
-{ "index" : { "_index" : "persons", "_id": "5360L000026752" } }
-{ "id": "4", "title": "eaUtlaborum velit cillum in occaecat", "firstname": "sint occaecatdeserunt euanim sed quis dolor", "middleName": "exercitation consequat eu", "surname": "eu id ea laboris inametad dolor anim Duis", "preferredFirstname": "veniamnulla commodo quiexercitation aliquip sunt adipisicing labore",      "preferredSurname": "in deserunt ut amet tempor", "ethinicity": "fugiatea mollit anim in nisi", "nationality": "exercitation nullasit do nulla sint eu", "placeOfBirth": "pariatur ex consectetur enim ad", "dateOfBirth": "adipisicing esse quis idnulla adipisicing ea elit non", "gender": "utculpa enim minim ullamco ad", "identification": [ { "identificationType": "fugiat nulla deserunt veniam", "value": "proident eiusmod quis pariatur", "originalDocumentSeen": true, "linkToDocument": "consectetur non ipsumanim ipsum dolor non" } ], "personTypes": [ "occaecat elit et sitin amet nisi est labore" ], "IsPersonCautionaryAlert": true, "IsTenureCautionaryAlert": true, "tenures": [ { "id": "dolor ullamco dolore deserunt", "type": "sit ut mollit magnanulla commodo dolor", "startDate": "officia exercitation irure deserunt", "endDate": "ullamco laboreautedeserunt Ut exercitation", "assetFullAddress": "sit laboris idpariatur amet est consequat" } ] }
+{ "index" : { "_index" : "persons", "_id": "5360L000026753" } }
+{ "id": "6f22e9ae3e8a4e0eaf46db02eb87f8e6", "title": "Mrs", "firstname": "Julie", "middlename": "", "surname": "Evans", "preferredFirstname": "string", "preferredSurname": "string", "ethinicity": "Christian", "nationality": "Canadian", "placeOfBirth": "Toronto", "dateOfBirth": "1990-02-19", "gender": "F", "identification": [{"identificationType": "NI", "value": "1234A", "originalDocumentSeen": true, "linkToDocument": "string" }], "personTypes": [ "Housing Officer" ], "IsPersonCautionaryAlert": true, "IsTenureCautionaryAlert": true, "tenures": [{ "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6", "type": "secure", "startDate": "string", "endDate": "string", "assetFullAddress": "Apartment 22, 18 G road, SW11"}]}
